// MochiR.Client/src/lib/api/types.ts
export interface components {
	schemas: {
// ...
		SelfProfileDto: {
			id: string;
			userName: string | null;
			displayName: string | null;
			email: string | null;
			emailConfirmed: boolean;
			phoneNumber: string | null;
			phoneNumberConfirmed: boolean;
			avatarUrl: string | null;
			twoFactorEnabled: boolean;
			lockoutEnabled: boolean;
			/** Format: date-time */
			lockoutEnd: string | null;
			/** Format: date-time */
			createdAtUtc: string;
			/** Format: int32 */
			followersCount: number;
			/** Format: int32 */
			followingCount: number;
		} | null;
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}